Refactored and modulized the bot. Add custom logger, and test.
- updated package.json to reflect to updated repository on abzico on
github (as transferred from haxpor to abzico)
- there is options object that can be at the time of starting the bot
- logger.js wraps around console object but provides flexibility to
disable or enable logging
- solution to reliably send message is added in mediator.js but not yet
integrated into logic.js. After we updated text to pre node for replying
message, we have to switch to another conversation by clicking on it (in
this case the best option is to switch to file transfer) then switch
back to previously conversation item, then click on the send button.
This way we can send message reliably every single time.