Add detecting new muted message as well as muted at the same time
Use JS selector combined with querySelector() to get the first element
that found to work on it.

Important note at the moment is that, we didn't have mechanism to keep
track of latest received message from muted conversation yet thus we
always return 1 as number of new message when we knew it's muted
message.