Add re-login ability
This in turned refactored code for 3 callbacks used for openURL()
function in index.js.

The scenario of re-login is needed when user logs in elsewhere on Web or
WeChat desktop app. It thus invalidates the current session. We handle
it by invalidating the current states, and starting over via start()
with the previously supplied arguments by user.