Add listening to particular user/group chat to not get rid of msg notif
Also
- Fixed left-over code to return properly in mediator.js
- Removed un-needed comment of function

Even we added this, there is a major issue now that WeChat's user id
changes every time you sign in so it's pretty hard to do mapping. Might
need to map from display name instead, but it's so volatile and not
practical to rely on it.

We will find solution later.